Moved cluster interface to kubernetes
